apply code review fixes
Critical fixes:
- audio_tagger: guard against eyed3.load() returning None in get_mp3()
- audio_extractor: guard against clip.audio being None before write_audiofile()
- audio_extractor: change AttributeError to FileNotFoundError for missing file
- audio_extractor/tagger: fix str=None default params to str|None=None

Type annotation improvements:
- process_class: add -> None to process_file abstract method
- utils: replace typing.Tuple/List with built-in tuple/list and Collection
- utils: add missing full_path: bool annotation; fix docstring
- audio_normalizer: add -> None to process_file
- audio_tagger: add eyed3.mp3 import; annotate mp3_file as Mp3AudioFile|None
- files_processor: replace typing imports with collections.abc; str|Path for file_dir

Code quality:
- constants: add module docstring; change VID_EXTS/AUDIO_EXTS to frozenset
- audio_extractor/normalizer: store audio_dir/normalized_dir as Path; use / operator
- files_processor: Collection[str] for ext_list
- normalize_audios_from_dir/files_processor/audio_normalizer: --dBFS type=float

Test improvements:
- test_audio_extractor: shared mock_clip_env fixture; FileNotFoundError;
  Path-based assertions; new test for no-audio-track ValueError
- test_audio_normalizer: Path-based assertions; top-level imports
- test_audio_tagger: top-level imports; new test for get_mp3 None return
- test_utils/files_processor: remove unused pytest import; top-level imports

+++ src/normalize_audios_from_dir.py
@@ -14,7 +14,7 @@ if __name__ == "__main__":
         help="Directory containing audio files to normalize.",
     )
     parser.add_argument(
-        "--dBFS", type=int, default=-30, help="Target dBFS. Defaults to -30."
+        "--dBFS", type=float, default=-30, help="Target dBFS. Defaults to -30."
     )
     args = parser.parse_args()
     process_all_files(args.dir, AUDIO_EXTS, AudioNormalizer, target_dbfs=args.dBFS)
